Removing path.py from REQUIREMENTS.
I really love path.py, I really do. The problem with it is that it makes you go
out of your way to install it system-wide before you can do anything else. This
isn't worth the tradeoff in my opinion. I'd like to use a consistent API
everywhere (in my fabfiles, in my django project settings, etc.) so,
unfortunately, I'm removing it in favor of os.path.

Yes, I know it sucks, but it's more consistent this way. :(