Adding django-sentry 1.6.6.1 to REQUIREMENTS.
django-sentry is the coolest way to track errors on your Django apps. It's made
by the Disqus team, and works really well (and has a sleek interface). Using it
makes everything else seem like a total piece of shit, so it's part of our
Django 1.3 stack.

If something better comes along in the future then I'll remove it and swap it
out, but for now this thing is king.