Adding `djkombu` to settings.dev.
djkombu is a ghetto queue broker for celery so that we can use it with django
without relying on an external queue service like rabbitmq or redis.

This is the best way for us to do it for development, since we can still use
queued tasks in our code and test them without managing external shit.